rpiserver/pyproject.toml
[project]
name = "gscontroller"
version = "1.0.0"
description = "Galicic Scientific Autosampler and Syringe Pump Controller"
dependencies = [
    "flask",
    "pyserial"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"